将资源上传路径改为相对路径，并优化了资源上传相关代码；增加了README文档

+++ src/main/java/com/tmyx/backend/controller/ArticleController.java
@@ -13,6 +13,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -46,10 +47,12 @@ public class ArticleController {
     public Result preGenerateId() {
         // 生成文章id
         String articleId = UUID.randomUUID().toString().replace("-", "");
+        // 获取绝对路径
+        File baseDir = new File(baseUploadPath).getAbsoluteFile();
         // 拼接文章保存路径
         String subPath = "/article/" + articleId + "/images/";
-        File folder = new File(baseUploadPath + subPath);
-
+        File folder = new File(baseDir + subPath);
+        // 如果目录不存在则创建
         if (!folder.exists()) {
             boolean success = folder.mkdirs();
             if (!success) {
@@ -64,21 +67,24 @@ public class ArticleController {
     public Result uploadImage(@RequestParam("file") MultipartFile file,
                               @RequestParam("articleId") String articleId) {
         if (file.isEmpty()) return Result.error("文件不能为空");
+        // 获取绝对路径
+        File baseDir = new File(baseUploadPath).getAbsoluteFile();
 
         try {
             // 确定文章的存储路径
-            String subPath = "article/" + articleId + "/images/";
-            File folder = new File(baseUploadPath + subPath);
+            String subPath = "/article/" + articleId + "/images/";
+            File folder = new File(baseDir + subPath);
+            // 如果目录不存在则创建
             if (!folder.exists()) {
                 folder.mkdirs();
             }
 
             String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
             File dest = new File(folder, fileName);
-            file.transferTo(dest);
+            file.transferTo(dest.getAbsoluteFile());
 
             // 拼接路径
-            String fileUrl = "/files/" + subPath + fileName;
+            String fileUrl = "/files" + subPath + fileName;
             return Result.success(fileUrl);
         } catch (IOException e) {
             e.printStackTrace();
@@ -89,21 +95,24 @@ public class ArticleController {
     // 保存文章
     @PostMapping("/save")
     public Result saveArticle(@RequestBody Article article, @RequestAttribute Integer userId) {
-        // 确定文件存储相对路径以及文件名
-        // article/{articleId}/index.md
-        String relativeFolder = "/article/" + article.getId() + "/";
-        String fileName = "index.md";
         // 获取绝对路径
-        File folder = new File(baseUploadPath + relativeFolder);
+        File baseDir = new File(baseUploadPath).getAbsoluteFile();
+        String absolutePath = baseDir.getAbsolutePath();
+        // 确定文件存储路径以及文件名
+        // article/{articleId}/index.md
+        String subPath = "/article/" + article.getId();
+        File folder = new File(baseDir + subPath);
+        // 如果目录不存在则创建
         if (!folder.exists()) folder.mkdirs();
         // 清除无用图片
         if (article.getContent() != null) {
-            FileUtil.cleanOrphanImages(article.getContent(), article.getId(), baseUploadPath);
+            FileUtil.cleanOrphanImages(article.getContent(), article.getId(), absolutePath);
         }
         // 将前端传来的content写入markdown文件中
         if (article.getContent() != null) {
-            try (FileWriter writer = new FileWriter(new File(folder, fileName))) {
-                writer.write(article.getContent());
+            try {
+                File mdFile = new File(folder, "index.md");
+                Files.writeString(mdFile.toPath(), article.getContent(), StandardCharsets.UTF_8);
             } catch (IOException e) {
                 e.printStackTrace();
                 return Result.error("内容文件保存失败");
@@ -114,7 +123,7 @@ public class ArticleController {
         // 组装文章信息
         article.setId(article.getId());
         article.setUpId(userId);
-        article.setContentUrl("/files" + relativeFolder + fileName);
+        article.setContentUrl("/files" + subPath + "/index.md");
         article.setUploadTime(LocalDateTime.now());
         // 检查文章是否存在
         Article existArticle = articleMapper.findById(article.getId());
@@ -133,10 +142,13 @@ public class ArticleController {
         if (article == null) return Result.error("文章不存在");
 
         String contentUrl = article.getContentUrl();
+        // 获取绝对路径
+        File baseDir = new File(baseUploadPath).getAbsoluteFile();
+        String absolutePath = baseDir.getAbsolutePath();
 
         if (contentUrl != null && !contentUrl.isEmpty()) {
             String relativePath = contentUrl.replace("/files/", "");
-            Path fullPath = Paths.get(baseUploadPath, relativePath);
+            Path fullPath = Paths.get(absolutePath, relativePath);
             File mdFile = fullPath.toFile();
 
             if (mdFile.exists()) {
@@ -159,16 +171,23 @@ public class ArticleController {
         // 根据文章id获取文章信息
         Article article = articleMapper.findById(id);
         if (article == null) return Result.error("文章不存在");
-        // 删除数据库记录
-        int rows = articleMapper.deleteById(id);
+
+        Path bassRoot = Paths.get(baseUploadPath).toAbsolutePath().normalize();
         // 获取文件路径
         String fileUrl = article.getContentUrl();
         // 去除映射路径的/files前缀，再去掉markdown文件的文件名，得到相对路径
         String relativeDirPath = fileUrl.replace("/files/", "").replace("index.md", "");
+        // 获取绝对路径
+        Path fullPath = bassRoot.resolve(relativeDirPath).normalize();
         // 拼接物理路径并删除
-        File articleFolder = new File(baseUploadPath + relativeDirPath);
-        FileUtil.deleteDirectory(articleFolder);
-
+        File articleFolder = fullPath.toFile();
+        if (articleFolder.exists() && articleFolder.isDirectory()) {
+            FileUtil.deleteDirectory(articleFolder);
+        } else {
+            System.out.println("目录不存在");
+        }
+        // 删除数据库记录
+        articleMapper.deleteById(id);
         return Result.success();
     }
